updated format
The key and pair sizes now take up exactly 8 bytes combined and are the
first bytes of the format. The next block is the value, then the key.
The thinking is that the value will likely have the most need to be
memory-aligned and starting at the 8th byte could provide performance
gains for user-defined data structures. The key is likely to be textual
and alignment is less likely a concern.